move error.ts to common folder
